This commit changes some of the workshop chat outputs by turning ** into italics and **** into bold.
I also tried doing the same to bottom stack and project window.
They work for the preview, but when appended the italics and bold disappear.

still it is good for better readability.
Didn't see any new bugs.

Finally the start.bat file was changed to create
a new venv into the directory of the software
so that no packages may conflict during usage.